Version update to '4.3.0'
Minor cosmetic changes in the output

Solved a bug in the compilation system of the 'common_utilities'
that prevent the Dimemas trace to create nanosecond traces.

Using negative values in the 'modules information' field
of the CFG file now indicated the simulator that all
CPU bursts inside this module will change its duration to
absoulte value of the number indicated. Required to
perform 'forced balancing' simulations.