New way to compute the collective FANs, taking into account the number of partners at different stages: - Number of tasks per node - Number of nodes per machine - Number of machines in WAN
Solved a bug when loading the collectives fine tuning
using an external file. It loaded the values mixing the
internal network and the WAN.

Solved a bug when reading blank lines of a trace with
CRLF line terminators

Added a warning when wrong records (TaskIds out of bounds
found in the trace

Add missing include of 'list.h' to 'simulator.c' file that
caused a segmentation fault in linux-ia64 machines.

Add missing information about collective communications ids
to the 'extern_comm_model.h' header file