Version update 5.0.0
Added intra-node contention simulation. This contention
is modelled in the same way as the intra-node contention,
using # of buses, # of input links and # of output links.
It requires adding these three parameters to the configuration
file ('node information' has changed), the GUI and also
the DimemasUpdateCFG, to translate old CFG files no the
newer version.

Due to the introduction of these new feature and the
modificacion of the configuration file, the major version
has been increased.

'prv2dim' did not translate the MPI_Rsend primitives. Now
they are translated as regular MPI_Send.