The management of parameter '--lat' was incorrect when using the new configuration files
Added a barrier synchronization primitive inside each
'MPI_Init' call when translating it. Also added the
parameter '-s' in the translator to avoid the generation
of this synchronization point.

Added the 'ezOptionsParser' header to manage the parameters
in 'prv2dim'. It is an easy solution to deal with
positional parameters and flags, but does not check the
types at all. It is planned to add it to the rest of the
package.

Added the 'bsc_utils' to check if a string or char pointer
codifies a double or a long int, and the corresponding
translators. This class could be extended. Other option is
to use the C++11 'strtod' and similars.