Added support for 'cyclic simulations', i.e. simulations of Paraver traces that were previously simulated. So far, we use a fixed events (from 9 to 14, SHOULD BE CHANGED!) to indicate the different communication endpoints in the first translation (original Paraver trace). Following translations will use this events to generate the endpoints, instead of regular communications.
TO CHECK: in following translations, the detection of 'this is
a Paraver trace generated by Dimemas' is done when we detect the
first 'special event'. That shouldn't be done like this (a regular
communication may appear before an event, and crash the translation).
A good approach will be using the 'TraceMetadata', and indicate in
the Paraver trace that it was generated using Dimemas!!