Added support for non blocking collectives.
Prv2dim has been changed in order to recognise the new operations. Now the global operations record has a new field. This new field has been placed on the right-hand side and indicates if this global operations is blocking (1), if is non blocking (0) or if it is a wait for a global operation (2). For the moment the matching between waitings and non blocking collectives is done by a FIFO.

From the Dimemas side, the model for the non blocking collectives is like the blocking ones but performed in background. Once a wait is achieved it pais the startup time and look for if the collective is done, in this case it returns immediatelly, if not it waits until the operation is done.

Change-Id: I1fbd5a9fdac02d050e6171d78284dfddc430b8ae